// src/lib/types.ts
export interface VulnmapProject {
  name: string;
  id: string;
  created: string;
  origin: string;
  type: string;
  branch: string | null;
  status: 'inactive' | 'active';
}

export interface Org {
  name: string;
  id: string;
  slug: string;
  url: string;
  group: {
    name: string;
    id: string;
  };
}

export interface RESTProjectData {
  attributes: RESTProjectsAttributes;
  id: string;
  relationships: RESTProjectsRelationships;
  type: string;
}

export interface RESTProjectsAttributes {
  businessCriticality: string;
  created: string;
  environment: string;
  lifecycle: string;
  name: string;
  origin: string;
  status: 'inactive' | 'active';
  tags: unknown;
  targetReference: string | null;
  type: string;
}

export interface RESTProjectsRelationships {
  importingUser: RESTProjectsRelashionshipData;
  org: RESTProjectsRelashionshipData;
  owner: RESTProjectsRelashionshipData;
  target: RESTProjectsRelashionshipData;
}

export interface RESTProjectsRelashionshipData {
  data: {
    id: string;
    type: string;
  };
  links: {
    first?: string;
    last?: string;
    next?: string;
    prev?: string;
    related?: string;
    self?: string;
  };
  meta: unknown;
}

export interface VulnmapTargetRelationships {
  org: {
    data: {
      id: string;
      type: string;
    };
    links: {
      first?: string;
      last?: string;
      next?: string;
      prev?: string;
      related?: string;
      self?: string;
    };
    meta: unknown;
  };
}

export interface RepoMetaData {
  branch: string;
  cloneUrl: string;
  sshUrl: string;
  archived: boolean;
}

export interface VulnmapTarget {
  attributes: {
    displayName: string;
    isPrivate: boolean;
    origin: string;
    remoteUrl: string | null;
  };
  id: string;
  relationships: VulnmapTargetRelationships;
  type: string;
}

export interface RESTTargetResponse {
  data: VulnmapTarget[];
  jsonapi: {
    version: string;
  };
  links: {
    first?: string;
    last?: string;
    next?: string;
    prev?: string;
    related?: string;
    self?: string;
  };
}